+++ nix/haskell-project.nix
@@ -58,14 +58,31 @@ in
                   hpack
                 ''
             else root;
+          setSrc = name: root: pkg: hlib.overrideSrc {
+                src =
+                  lib.cleanSourceWith {
+                    name = "source-${name}-${pkg.version}";
+                    src =
+                      # NOTE: Even though cabal2nix does run hpack automatically,
+                      # buildFromCabalSdist does not. So we must run hpack ourselves at
+                      # the original source level.
+                      realiseHpack name root;
+                  };
+              } pkg;
         in
         lib.mapAttrs
-          (name: value:
-            # NOTE: Even though cabal2nix does run hpack automatically,
-            # buildFromCabalSdist does not. So we must run hpack ourselves at
-            # the original source level.
-            let root = realiseHpack name value.root;
-            in fromSdist (self.callCabal2nix name root { })
+          (name: pkgCfg:
+            lib.pipe pkgCfg
+            [
+              (pkgCfg: self.callCabal2nix name pkgCfg.root { })
+
+              # Avoid rebuilding because of changes in parent directories
+              (setSrc name pkgCfg.root)
+
+              # Make sure all files we use are included in the sdist, as a check
+              # for release-worthiness.
+              fromSdist
+            ]
           )
           config.packages;