+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   - #92: Add `devShell.mkShellArgs` to pass custom arguments to `mkShell`
   - #111: Add `devShell.extraLibraries` to add custom Haskell libraries to the devshell.
   - #100: `source-overrides` option now supports specifying Hackage versions.
+  - #114: Prevent Nix rebuilds of packages in sub-directories when parent contents change.
 - API changes
     - #37: Group `buildTools` (renamed to `tools`), `hlsCheck` and `hlintCheck` under the `devShell` submodule option; and allow disabling them all using `devShell.enable = false;` (useful if you want haskell-flake to produce just the package outputs).
     - #64: Remove hlintCheck (use [treefmt-nix](https://github.com/numtide/treefmt-nix#flake-parts) instead)